Delete checked in binary, fix test renaming inside go.mod and *.go files
I'm assuming the binary wasn't supposed to be there.
